Fix notifiations; Verify portal parsing on local
There were a bunch of SSL cert issues trying to run everything on https
by IP, so I switched it all back to localhost and verified it that way.

Everything seems to work nicely.

I was also able to fix the notifications not cleaning themselves up, and
also prevent auth routes from trying to load when the session ID isn't
present. This cleaned a lot of console error messages.